+++ src/pages/Home.js
@@ -1,127 +1,223 @@
-import React, { useState, Fragment, useCallback } from "react";
-import {
-  GoogleMap,
-  useJsApiLoader,
-  Marker,
-  InfoBox,
-} from "@react-google-maps/api";
-import { useLocation } from "react-router-dom";
+import React, { useState, useEffect, Fragment, useRef } from "react";
+import axios from "axios";
 
+import Map from "../components/Home/Map_Home";
+import Searchbar from "../components/Home/Searchbar_Home";
+import Itemlist from "../components/Home/Itemlist_Home";
 import Info from "../components/Home/Info_Home";
 
-const containerStyle = {
-  width: "100%",
-  height: "100vh",
-  position: "absolute",
-};
+const API_URL_GEO =
+  "https://asianfood.heguangyu.net/return_location_geo_react.php";
+
+const API_URL_TYPE =
+  "https://asianfood.heguangyu.net/return_location_type_react.php";
+
+const API_URL_REVIEW_DETAILS =
+  "https://asianfood.heguangyu.net/return_location_review_details_react.php";
+
+const API_URL_NAME_OFTYPE =
+  "https://asianfood.heguangyu.net/return_location_name_oftype_react.php";
+
+const API_URL_NAME_ANDTYPE =
+  "https://asianfood.heguangyu.net/return_location_name_andtype_react.php";
 
 function Home() {
-  let mapOptions = {
-    zoom: 13,
-    zoomControl: true,
-    scaleControl: false,
-    fullscreenControl: false,
-    mapTypeControl: false,
-    streetViewControl: false,
-    gestureHandling: "greedy",
-    maxZoom: 15 + 3,
-    mapId: "e04d39f76af137b0",
-  };
+  //ANCHOR create ref for searchInput
+  const searchInput = useRef();
 
-  if (window.matchMedia("(prefers-color-scheme: dark)").matches) {
-    console.log("dark mode on");
-    mapOptions.mapId = "da23336c10c105b2";
-  }
-
-  let marker_pos = "";
-  let info_name = "";
   let info_state = false;
-  let type = "";
-  let review = 0;
-  let review_details = "";
 
-  const location = useLocation();
-  if (location.state === null) {
-    marker_pos = {
-      lat: 52.517674728732054,
-      lng: 13.393789389208452,
-    };
-    info_name = "Asian Food in Berlin";
-    info_state = false;
-    type = "";
-    review = 0;
-    review_details = "";
-  } else {
-    marker_pos = {
-      lat: parseFloat(location.state.lat),
-      lng: parseFloat(location.state.lng),
-    };
-    info_name = location.state.geo_name;
-    info_state = true;
-    type = location.state.type_name;
-    review = parseInt(location.state.review_points);
-    review_details = location.state.review_text;
+  //SECTION do when input has changed
+  const handleInputChange = async (event) => {
+    //ANCHOR open item listview and say loading
+    change_listview_state(true);
+    change_listitem(["loading..."]);
+
+    //ANCHOR query from input change
+    let query = event.target.value;
+
+    //ANCHOR a query list for potential searching through type
+    let kor_name_list = ["k", "K", "ko", "Ko", "kor", "Kor"];
+    let chi_name_list = ["c", "C", "ch", "Ch", "chi", "Chi"];
+    let jap_name_list = ["j", "J", "ja", "Ja", "jap", "Jap"];
+    let vie_name_list = ["v", "V", "vi", "Vi", "vie", "Vie"];
+
+    if (query && query.length > 0) {
+      //ANCHOR show option for searching by type
+      let names = [];
+      if (kor_name_list.indexOf(query) >= 0) {
+        names = [...names, "Korean Restaurants..."];
+      } else if (chi_name_list.indexOf(query) >= 0) {
+        names = [...names, "Chinese Restaurants..."];
+      } else if (jap_name_list.indexOf(query) >= 0) {
+        names = [...names, "Japanese Restaurants..."];
+      } else if (vie_name_list.indexOf(query) >= 0) {
+        names = [...names, "Vietnamese Restaurants..."];
+      }
+
+      //ANCHOR require data from php
+      //ANCHOR display both name and review
+      let results = [];
+      await axios.get(`${API_URL_NAME_ANDTYPE}?n=${query}`).then(({ data }) => {
+        results = data.split(",");
+      });
+      //ANCHOR handle data into arrays
+      for (let i = 0; i < results.length - 1; i++) {
+        if (results[i] === undefined) {
+        } else {
+          names = [...names, results[i]];
+        }
+      }
+      //ANCHOR display arrays in list
+      change_listitem(names);
+    } else if (!query) {
+      //ANCHOR if nothing there, close the list view
+      change_listview_state(false);
+    }
+  };
+  //!SECTION
+
+  //SECTION control listview component state
+  const [listview_state, set_listview_state] = useState(false);
+  function change_listview_state(props) {
+    setTimeout(() => {
+      set_listview_state(props);
+    }, 100);
+  }
+  //!SECTION
+
+  //SECTION set list items
+  const [listitem, set_listitem] = useState([]);
+  function change_listitem(props) {
+    setTimeout(() => {
+      set_listitem(props);
+    }, 100);
+  }
+  //!SECTION
+
+  //SECTION add marker on Home.js when clicking items from list
+  //ANCHOR using useNavigate to pass props
+  //const navigate = useNavigate();
+  let geo_ini = {
+    lat: 0,
+    lng: 0,
+    geo_name: "",
+    type_name: "",
+    review_points: "",
+    review_text: "",
+  };
+  const [geo, set_geo] = useState(geo_ini);
+  function change_geo(props) {
+    setTimeout(() => {
+      set_geo(props);
+    }, 100);
   }
 
-  const { isLoaded } = useJsApiLoader({
-    id: "google-map-script",
-    googleMapsApiKey: "AIzaSyD7eBZcwtUqmgeuYkWPjDK8d0uvU3Q1WrE&language=en",
-  });
+  const handelClick_item = async (query) => {
+    //ANCHOR initial a result container
+    let results = [],
+      names = "",
+      type = "",
+      review_details = "";
 
-  // eslint-disable-next-line
-  const [map, setMap] = useState(null);
+    //ANCHOR say loading when querying data form server
+    change_listitem(["loading..."]);
 
-  const onLoad = useCallback(function callback(map) {
-    setMap(map);
-  }, []);
+    //SECTION query data using type
+    let type_query = "";
+    if (query === "Korean Restaurants...") {
+      type_query = "ko";
+    } else if (query === "Chinese Restaurants...") {
+      type_query = "ch";
+    } else if (query === "Japanese Restaurants...") {
+      type_query = "ja";
+    } else if (query === "Vietnamese Restaurants...") {
+      type_query = "vi";
+    } else {
+    }
+    if (type_query === "") {
+    } else {
+      await axios
+        .get(`${API_URL_NAME_OFTYPE}?n=${type_query}`)
+        .then(({ data }) => {
+          results = data.split(",");
+        });
+      for (let i = 0; i < results.length - 1; i++) {
+        if (results[i] === undefined) {
+        } else {
+          names = [...names, results[i]];
+        }
+      }
+      change_listitem(names);
+      return false;
+    }
+    //!SECTION
 
-  const onUnmount = useCallback(function callback(map) {
-    setMap(null);
-  }, []);
+    //ANCHOR separate name and review, only use name in query here and review later
+    let name_query = query.split(";")[0];
+    if (name_query && name_query.length > 0) {
+      await axios.get(`${API_URL_GEO}?n=${name_query}`).then(({ data }) => {
+        results = data.split(",");
+      });
+      for (let i = 0; i < results.length; i++) {
+        if (results[i] === undefined) {
+        } else {
+          names = [...names, results[i]];
+        }
+      }
+      await axios.get(`${API_URL_TYPE}?n=${name_query}`).then(({ data }) => {
+        type = data;
+      });
+      await axios
+        .get(`${API_URL_REVIEW_DETAILS}?n=${name_query}`)
+        .then(({ data }) => {
+          review_details = data;
+        });
+      //ANCHOR build a object accepted by Marker in Home.js
+      change_geo({
+        lat: names[0],
+        lng: names[1],
+        geo_name: name_query,
+        type_name: type,
+        review_points: query.split(";")[1],
+        review_text: review_details,
+      });
 
-  return isLoaded ? (
+      //ANCHOR navigate to home.js with geo object
+      //console.log(window.location.pathname);
+
+      change_listview_state(false); //ANCHOR close list when loaded
+
+      //navigate(window.location.pathname, { state: geo });
+
+      //ANCHOR clean searchbar input
+      searchInput.current.value = "";
+    } else if (!query) {
+    }
+  };
+  //!SECTION
+
+  return (
     <Fragment>
-      <GoogleMap
-        mapContainerStyle={containerStyle}
-        options={mapOptions}
-        center={marker_pos}
-        onLoad={() => onLoad}
-        onUnmount={() => onUnmount}
-      >
-        {/* Child components, such as markers, info windows, etc. */}
-        <Marker position={marker_pos} />
-        <InfoBox
-          position={marker_pos}
-          options={{ closeBoxURL: "", enableEventPropagation: true }}
-        >
-          <div
-          className="
-          p-2 w-24
-          shadow-lg rounded-lg
-          bg-white dark:bg-gray-500
-          "
-          style={{opacity: "0.8"}}
-          >
-            <div className="
-            text-base
-            dark:text-white 
-            ">
-              {info_name}
-            </div>
-          </div>
-        </InfoBox>
-      </GoogleMap>
+      <Searchbar
+        searchInput={searchInput}
+        handleInputChange={handleInputChange}
+      />
+      <Itemlist
+        listview_state={listview_state}
+        listitem={listitem}
+        handelClick_item={handelClick_item}
+      />
+      <Map geo={geo}></Map>
       <Info
-        info_state={info_state}
-        info_name={info_name}
-        type={type}
-        review={review}
-        review_details={review_details}
+        info_state={true}
+        info_name={geo.geo_name}
+        type={geo.type_name}
+        review={geo.review_points}
+        review_details={geo.review_text}
       ></Info>
     </Fragment>
-  ) : (
-    <></>
   );
 }
 
-export default Home;
+export default React.memo(Home);
